+++ app/controllers/api/comments_controller.rb
@@ -1,11 +1,11 @@
 class Api::CommentsController < ApplicationController
     def index
         @post = Post.find(params[:post_id])
-        @comments = @post.comments
+        @comments = @post.comments.includes(:user)  
   
-        render json: @comments
+        render 'api/comments/index'
       end
-      
+
     def create
       @comment = Comment.new(comment_params)
       @comment.user_id = current_user.id
