+++ app/views/api/comments/index.json.jbuilder
@@ -1,0 +1,7 @@
+@comments.each do |comment|
+    json.set! comment.id do
+        json.extract! comment, :id, :text, :user_id, :post_id, :created_at, :updated_at
+        json.username comment.user.username
+        json.userPhotoUrl comment.user.photo.attached? ? url_for(comment.user.photo) : nil
+    end
+end
